+++ workspaces/announcements/plugins/announcements-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-announcements-react
 
+## 0.28.0
+
+### Minor Changes
+
+- 28ac778: Backstage version bump to v1.53.0
+
+### Patch Changes
+
+- Updated dependencies [28ac778]
+  - @backstage-community/plugin-announcements-common@0.23.0
+
 ## 0.27.0
 
 ### Minor Changes